output: Extract file review change construction
File-review model construction now has separate selection and pagination
helpers. The remaining row segmentation still built ReviewChange records
inside the model builder.

Keeping row segmentation in the orchestration layer made that layer own
selection lookup, unselectable notes, gap flushing, and change totals.

This commit moves ReviewChange assembly into output.file_review_changes and
has the model builder delegate after deriving actionable selections.

Subsequent commits will add architecture coverage for the
change-construction boundary.
